Adicionando oo codigos de aula-4/ex-5 e aula-4/ex-6

+++ matheus-586/aula-4/ex-7/Program.cs
@@ -1,14 +1,14 @@
 ﻿// - Escreva um programa que leia três notas de um aluno, calcule a média e exiba uma mensagem indicando se ele foi aprovado (média ≥ 7), está em recuperação (5 ≤ média < 7) ou foi reprovado (média < 5). Usar funções.
 Console.WriteLine("escreva a primeira nota do aluno: ");
-int nota1 = int.Parse(Console.ReadLine());
+double nota1 = double.Parse(Console.ReadLine());
 
 Console.WriteLine("escreva a segunda nota do aluno: ");
-int nota2 = int.Parse(Console.ReadLine());
+double nota2 = double.Parse(Console.ReadLine());
 
 Console.WriteLine("escreva a terceira nota do aluno: ");
-int nota3 = int.Parse(Console.ReadLine());
+double nota3 = double.Parse(Console.ReadLine());
 
-int somatorio = nota1 + nota2 + nota3;
-int media = somatorio/3;
+double somatorio = nota1 + nota2 + nota3;
+double media = somatorio/3;
 
 Console.WriteLine("Sua média é: " + media);
